% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3io-package.R
\docType{package}
\name{s3io-package}
\alias{s3io}
\alias{s3io-package}
\title{s3io: AWS S3 I/O Utilities}
\description{
Permits some forms of (in)direct reading and writing from/to files on AWS S3 buckets.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/mmuurr/s3io/issues}
}

}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com}

}
\keyword{internal}
