% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3io_read.R
\name{s3io_read}
\alias{s3io_read}
\title{Read a file in from AWS S3}
\usage{
s3io_read(
  bucket,
  key,
  readfun,
  ...,
  .localfile = fs::file_temp(),
  .rm_localfile = TRUE,
  .opts = NULL
)
}
\arguments{
\item{bucket}{<chr(1)> AWS S3 bucket name.}

\item{key}{<chr(1)> AWS S3 object key.}

\item{readfun}{<function> A read function where the first argument is a filepath.
The function's signature should look like \code{readfun(file, ...)}.}

\item{...}{Additional arguments forwarded on to \code{readfun}.}

\item{.localfile}{<chr(1)> The local filepath of the downloaded file.}

\item{.rm_localfile}{<lgl(1)> Should the local filepath be deleted once it has been read?
If \code{rm_localfile = TRUE}, then the read operation should read the entire file into memory.}

\item{.opts}{<dict> AWS CLI additional --opts, if not the default (from {awscli2}).}
}
\value{
The returned value from \code{readfun}.
}
\description{
Read a file in from AWS S3
}
\examples{
\dontrun{
  s3io_read("mybucket", "my/object/key.csv", readr::read_csv, col_names = TRUE)
}
}
