% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_file_remove.R
\name{try_file_remove}
\alias{try_file_remove}
\title{Try to remove a file}
\usage{
try_file_remove(filepath)
}
\value{
\code{TRUE} if the removal was successfuly, \code{FALSE} otherwise.
}
\description{
Tries to safely remove a file by capturing and ignoring any errors.
}
