% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3io_write.R
\name{s3io_write}
\alias{s3io_write}
\title{Write a file to AWS S3}
\usage{
s3io_write(obj, bucket, key, writefun, ..., localfile = tempfile(),
  rm_localfile = TRUE)
}
\arguments{
\item{obj}{The object to write.}

\item{bucket}{[string] AWS S3 bucket name.}

\item{key}{[string] AWS S3 object key.}

\item{writefun}{[function] A write function where the first argument is the object to write and the second argument is a filepath.
The function's signature should look like \code{writefun(obj, file, ...)}.}

\item{...}{Additional arguments passed on to \code{writefun}.}

\item{localfile}{[string] The local filepath for the initial write-to-disk.}

\item{rm_localfile}{[boolean] Remove \code{localfile} once the copy-to-S3 is complete?}
}
\value{
The returned value from \code{writefun}.
}
\examples{
\dontrun{
  s3io_write(iris, "mybucket", "my/object/key.csv", readr::write_csv, col_names = TRUE)
}
}
